pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-asr"
version = "0.0.6"
description = "Qwen-ASR python package"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
license = { text = "Apache-2.0" }
authors = [{ name = "Alibaba Qwen Team" }]

dependencies = [
  "torch",
  "torchaudio",
  "transformers==4.57.6",
  "accelerate==1.12.0",
  "qwen-omni-utils",
  "librosa",
  "soundfile",
  "openvino>=2026.0.0",
  "nncf>=3.0.0",
  "PySide6",
  "optimum-intel[openvino]>=1.27.0",
  "openvino-genai>=2026.0.0.0",
]

[project.urls]
Homepage = "https://github.com/Qwen/Qwen3-ASR"
Repository = "https://github.com/Qwen/Qwen3-ASR"

[project.scripts]
qwen-asr-demo = "qwen_asr.cli.demo:main"
qwen-asr-demo-streaming = "qwen_asr.cli.demo_streaming:main"
qwen-asr-serve = "qwen_asr.cli.serve:main"

[tool.setuptools]
packages = { find = { where = ["."] , include = ["qwen_asr*"] } }
include-package-data = true

[tool.setuptools.package-data]
qwen_tts = ["py.typed", "**/*.dict"]

[tool.uv]
environments = ["sys_platform == 'win32'"]

[[tool.uv.index]]
name = "pytorch-xpu"
url = "https://download.pytorch.org/whl/xpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-xpu" }]
torchaudio = [{ index = "pytorch-xpu" }]
